Refactor WalletConnect integration: improve type safety, consolidate schemas, and remove unused types
